feat: Handle not found news item in NewsDetailsPage
The code changes in this commit add a check for a not found news item in the NewsDetailsPage component. If the news item is not found, the `notFound()` function from the `next/navigation` module is called. This ensures that the user is redirected to a not found page when accessing a non-existent news item.